Change notification scheme on get/set
Before, the scheme was:

  - for push, (rank of producer, ITEM_NOT_OFFSET)
  - for pull, (ITEM_NOT_OFFSET, 1)

where pull was different because of gaspi_read_notify interface.
Now, both versions have (modified_item_id, 1),
where modified_item_id is item_id, shifted 1 to the left, and OR'ed
with 0x1 - this guarantees that we won't wait for notification id 0,
which on GASPI behaves differently; and the result is truncated to
the size of gaspi_notification_id_t.

(Tip: use --word-diff for this commit)

Changes to be committed:
	modified:   src/pull.c
	modified:   src/push.c