Expand bitset to allow larger groups of processes
The per-item metadata structure has to keep information
about who wants to consume an item
(at least for the push implementation), and it did so
using a bitset represented by an uint64_t - which limited
the number of processes to 64, 1 per bit.

With this commit, that field is now an array with
a configurable length (by a macro at compile-time),
and the appropriate changes were done to the implementations.
(Namely, removed the consumers information on pull, where it
was not currently needed)

Changes to be committed:
	modified:   src/gssp.c
	modified:   src/gssp_internal.h
	modified:   src/pull.c
	modified:   src/push.c
	modified:   src/local.c