Fix shortcircuiting behaviour of gssp_set
When a gssp_set operation is invoked on the "Push" implementation, it is
necessary to consult which ranks expect the new item and push to them.
In terms of code, this is a loop that checked each bit on the consumer
bitset. With the new bitset structure (array of uints), the loop
condition has to change so that we continue in the loop even if the current
uint has no remaining bits set (because other uint could have more
consumers).

Changes to be committed:
	modified:   src/push.c