Fix library not able for 1 process usage
  + Remove UB from tests when running with 1 process

The library was erroring out on the gaspi_atomic_fetch_add
because that is the only thing which depends on a master rank,
which I set to 1 in the past to check if there where any hidden logic
dependencies on wheter 0 is the master.

The UB was having VLAs (yes, on the C++ code...) with 0-length.

Changes to be committed:
	modified:   src/gssp_internal.h
	modified:   tests/no_gssp_instance.cpp
	modified:   tests/slack_tests.cpp