Mitigate performance of registering consumers
  + Add app example to test for it

The problem is that, despite the metadata segment being unordered, the
consuming list is ordered, and so instead of looping over both, I can
loop over the metadata segment and use a binary search in the consuming
list. This cuts the time from 1400s to 90s, on my PC, in the added
example (exp.c.bak2).

(tip: use -w option when viewing this commit)

Changes to be committed:
	modified:   src/gssp.c
	new file:   apps/exp.c.bak2