Relax assertion checks on wait_or_die_limited
  + Change wait_or_die to limited version on pull

Due to mutithreading (when value is not the expected, because maybe
another thread with the same notification id got there first) or to an
unresolved bug (sucessfully wait for an item id which is not the one we
intended when the id is 65535), sometimes the library gives up. The
bandaid is to just not check those assertions.

(tip: use --word-diff option when viewing this commit)

Changes to be committed:
	modified:   pull.c
	modified:   aux/waitsome.c