Add extra argument to functions + macro
The extra argument is a pointer to a structure
that eventually will specify the context of
the Lapser instance.

To consider: why not a global array hold by the
Lapser internals, and the handle as an index into it?

The macro has the same name as the functions,
and essentialy fills the new extra argument with
a pointer to a future structure that will hold
configuration & context info.

In principle, any application code should be unaffected, unless they try
to use lapser_internal.h without also including lapser.h - in that case,
they will need to also include lapser.h