Add functions to manage lapser_ctx objects
Only creating and freeing, no customization allowed.
The structure is filled using the best available information, by
querying the allocated segments and already created queues. Because
GASPI creates 8 queues at the beginning, it may be necessary to close
some of them, and to add more advanced manipulation functions
(or just expose the struct internals, like GPI).

A thing to consider is the appearance of double pointers.
It may be advisable to start using opaque types for the handles,
to avoid that appearance (but this is more a consideration of aesthetics
and programmer UX than engineering).

Changes to be committed:
	modified:   ../include/multilapser.h
	modified:   CMakeLists.txt
	modified:   lapser.c
	new file:   lapser_ctx.c