Add tests to lapser_ctx structure
  + Add more former global variables to lapser_ctx
  + Change lapser_init to also initialize lapser_ctx if needed
  + Fix logic on finding segment ids, to consider lapser_ctx already
    filled but not yet allocated by lapser_init

The global variables added to the ctx structure were the lookup and
meta_lookup arrays, and the item_slot_size (already there, but not used).

The function lapser_init now receives a pointer to context pointer, so
it can overwrite the callers context pointer with an initialized
context. If NULL is passed, the default _lapser_global_ctx is used.

Changes to be committed:
	modified:   include/lapser.h
	modified:   include/multilapser.h
	modified:   src/lapser.c
	modified:   src/lapser_ctx.c
	modified:   src/lapser_internal.h
	modified:   src/local.c
	modified:   src/pull.c
	modified:   src/push.c
	modified:   tests/CMakeLists.txt
	modified:   tests/checksum_works.cpp
	new file:   tests/ctx_tests.cpp
	new file:   tests/simple_multilapser.cpp